requirements.txt
flake8==6.0.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
python-dotenv==0.21.1
PyYAML==6.0
SQLAlchemy==2.0.1
validators==0.20.0
WTForms==3.0.1
